update more dependencies
