+++ frontend/src/modules/Environments/components/EnvironmentDatasets.js
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types';
 import { Box } from '@mui/material';
-import EnvironmentSharedDatasets from './EnvironmentSharedDatasets';
-import EnvironmentOwnedDatasets from './EnvironmentOwnedDatasets';
+import { EnvironmentSharedDatasets } from './EnvironmentSharedDatasets';
+import { EnvironmentOwnedDatasets } from './EnvironmentOwnedDatasets';
 
-const EnvironmentDatasets = ({ environment }) => (
+export const EnvironmentDatasets = ({ environment }) => (
   <Box>
     <Box>
       <EnvironmentOwnedDatasets environment={environment} />
@@ -17,5 +17,3 @@ const EnvironmentDatasets = ({ environment }) => (
 EnvironmentDatasets.propTypes = {
   environment: PropTypes.object.isRequired
 };
-
-export default EnvironmentDatasets;
