+++ frontend/src/modules/Environments/components/EnvironmentFeatures.js
@@ -11,7 +11,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Label } from 'design';
 
-const EnvironmentFeatures = (props) => {
+export const EnvironmentFeatures = (props) => {
   const { environment, ...other } = props;
 
   return (
@@ -121,5 +121,3 @@ const EnvironmentFeatures = (props) => {
 EnvironmentFeatures.propTypes = {
   environment: PropTypes.object.isRequired
 };
-
-export default EnvironmentFeatures;
