+++ frontend/src/modules/Environments/components/EnvironmentSharedDatasets.js
@@ -27,7 +27,7 @@ import {
 import { SET_ERROR, useDispatch } from 'globalErrors';
 import { searchEnvironmentDataItems, useClient } from 'services';
 
-const EnvironmentOwnedDatasets = ({ environment }) => {
+export const EnvironmentSharedDatasets = ({ environment }) => {
   const client = useClient();
   const navigate = useNavigate();
   const dispatch = useDispatch();
@@ -176,8 +176,6 @@ const EnvironmentOwnedDatasets = ({ environment }) => {
   );
 };
 
-EnvironmentOwnedDatasets.propTypes = {
+EnvironmentSharedDatasets.propTypes = {
   environment: PropTypes.object.isRequired
 };
-
-export default EnvironmentOwnedDatasets;
