+++ frontend/src/modules/Environments/services/index.js
@@ -9,7 +9,6 @@ export * from './getEnvironmentAssumeRoleUrl';
 export * from './getPivotRoleExternalId';
 export * from './getPivotRoleName';
 export * from './getPivotRolePresignedUrl';
-export * from './getTrustAccount';
 export * from './inviteGroup';
 export * from './listAllEnvironmentConsumptionRoles';
 export * from './listAllEnvironmentGroups';
@@ -21,7 +20,6 @@ export * from './listEnvironmentMembers';
 export * from './listEnvironmentNetworks';
 export * from './listEnvironmentNotMembers';
 export * from './listEnvironmentPermissions';
-export * from './listEnvironments';
 export * from './listInvitedGroups';
 export * from './listOrganizationEnvironments';
 export * from './removeConsumptionRole';
