+++ frontend/src/services/graphql/Environments/index.js
@@ -1,0 +1,2 @@
+export * from './getTrustAccount';
+export * from './listEnvironments';
